+++ packages/camera/lib/camera.dart
@@ -586,4 +586,20 @@ class CameraController extends ValueNotifier<CameraValue> {
       await _eventSubscription?.cancel();
     }
   }
+
+  /// Flashlight switch:
+  /// enables flashlight if `value == true`,
+  /// disables flashlight if `value == false`
+  Future<void> flash(bool value) async {
+    try {
+      await _channel.invokeMapMethod<String, dynamic>(
+        'enableFlash',
+        <String, dynamic>{
+          'flash': value,
+        },
+      );
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+  }
 }